// src/models/protocolType.ts
/** Protocols and services that can be assigned to a network device. */
export enum ProtocolType {
  Rip = "rip",
  Ospf = "ospf",
  Vlan = "vlan",
  Stp = "stp",
  Rstp = "rstp",
  EtherChannel = "etherchannel",
  Dhcp = "dhcp",
  StaticRoute = "static-route",
  Acl = "acl",
  Nat = "nat",
  Pat = "pat"
}